Replace existing code with rewritten version
