+++ packages/core/src/computed.ts
@@ -1,43 +1,11 @@
-import { Atom } from "./atom";
-import { Observer } from "./observer";
+import { atom, Atom } from "./atom";
+import { effect } from "./effect";
 
-export class Computed<TValue> extends Atom<TValue> {
-  #observer: Observer;
-  #compute: () => TValue;
+export type Computed<TValue> = Atom<TValue>;
 
-  constructor(
-    compute: () => TValue,
-    { onNotify }: { onNotify?: () => void } = {}
-  ) {
-    super(compute());
+export function computed<TValue>(compute: () => TValue): Computed<TValue> {
+  const computed = atom(compute());
+  effect(() => computed(compute() as any));
 
-    this.readonly = true;
-
-    this.#compute = compute;
-    this.#observer = new Observer(this.#computeValue, {
-      onNotify,
-    });
-
-    this.#observer.exec();
-  }
-
-  #computeValue = () => {
-    this.readonly = false;
-    this.value = this.#compute();
-    this.readonly = true;
-  };
-
-  refresh = (compute?: () => TValue) => {
-    this.#compute = compute ?? this.#compute;
-    this.#observer.exec();
-  };
-
-  dispose = () => this.#observer.dispose();
-}
-
-export function computed<TValue>(
-  callback: () => TValue,
-  { onNotify }: { onNotify?: () => void } = {}
-) {
-  return new Computed(callback, { onNotify });
+  return computed;
 }
