+++ packages/core/src/effect.test.ts
@@ -1,23 +1,18 @@
 import { describe, expect, it } from "vitest";
 import { atom } from "./atom";
 import { effect } from "./effect";
+import { state } from "./state";
 
 describe("effect", () => {
-  it("should call when included atom updates", () => {
+  it("should call when included primitve updates", () => {
     const a1 = atom("abc");
     const a2 = atom(2);
-    const a3 = atom<string[]>([]);
 
     let effectCount = -1;
-    let anotherEffectCount = -1;
     effect(() => {
       a1.value;
       a2.value;
       effectCount++;
-    });
-    effect(() => {
-      a3.value[0];
-      anotherEffectCount++;
     });
 
     expect(effectCount).toBe(0);
@@ -25,11 +20,48 @@ describe("effect", () => {
     expect(effectCount).toBe(1);
     a2.value = 2.5;
     expect(effectCount).toBe(2);
-    expect(anotherEffectCount).toBe(0);
-    a3.value.push("bb");
-    expect(effectCount).toBe(2);
+  });
+
+  it("should call when included state updates", () => {
+    const a3 = state<string[]>([]);
+    let anotherEffectCount = -1;
+
+    effect(() => {
+      a3[0];
+      anotherEffectCount++;
+    });
+
+    a3.push("bb");
     expect(anotherEffectCount).toBe(1);
-    a3.value.push("cc");
-    expect(anotherEffectCount).toBe(1);
+  });
+
+  it("should call when deeply nested property updates", () => {
+    const a3 = state({ a: { b: { c: [{ d: 1 }] } } });
+    let deepEffectCount = -1;
+
+    effect(() => {
+      a3.a.b.c[0].d;
+      deepEffectCount++;
+    });
+
+    a3.a.b.c[0].d++;
+    expect(deepEffectCount).toBe(1);
+  });
+
+  it("should not call after disposed", () => {
+    const a1 = atom("abc");
+
+    let effectCount = -1;
+    const dispose = effect(() => {
+      a1.value;
+      effectCount++;
+    });
+
+    expect(effectCount).toBe(0);
+    a1.value = `xyz`;
+    expect(effectCount).toBe(1);
+    dispose();
+    a1.value = `zyx`;
+    expect(effectCount).toBe(1);
   });
 });
