+++ packages/core/src/ref.ts
@@ -1,13 +1,13 @@
-export const IS_REF = Symbol("IS_REF");
-
 class Ref<TValue> {
-  [IS_REF] = true;
-
   value: TValue;
 
   constructor(value: TValue) {
     this.value = value;
   }
+}
+
+export function isRef(value: any): value is Ref<any> {
+  return value instanceof Ref;
 }
 
 export function ref<TValue>(value: TValue) {
