+++ packages/core/src/types.ts
@@ -1,3 +1,3 @@
-import { Atom } from "./atom";
+import { State } from "./state";
 
-export type extractAtomValue<T> = T extends Atom<infer V> ? V : T;
+export type extractStateValue<T> = T extends State<infer V> ? V : T;
