+++ packages/core/src/util.ts
@@ -8,24 +8,6 @@ export function isClass(value: any) {
   );
 }
 
-export function canProxy(value: any) {
-  return (
-    value !== undefined &&
-    typeof value === "object" &&
-    (!isClass(value) ||
-      (!(value instanceof Map) &&
-        !(value instanceof Set) &&
-        !(value instanceof WeakMap) &&
-        !(value instanceof WeakSet) &&
-        !(value instanceof Error) &&
-        !(value instanceof Number) &&
-        !(value instanceof Date) &&
-        !(value instanceof String) &&
-        !(value instanceof RegExp) &&
-        !(value instanceof ArrayBuffer)))
-  );
-}
-
 export function arrayStartsWith(a: any[], b: any[]) {
   return b.every((bv, index) => bv === a[index]);
 }
