--- packages/core/src/v2/types.ts
@@ -1,3 +1,0 @@
-import { StateV2 } from "./state";
-
-export type extractStateV2Value<T> = T extends StateV2<infer V> ? V : T;
